refactor(tokens): complete colorblind audit/generator split per review
Round-2 review follow-up on #142:

- Relocate renderSvg, SVG layout constants, rgbToHex/svgEscape, and SVG_OUT
  from audit-colorblind.js into generate-colorblind-svg.js so the audit is a
  pure stdout+exit-code gate that owns no output file
- Extract shared titleCase into utils.js (used by the audit summary and the
  generator's renderSvg) instead of a cross-script reach-in
- Add a gamma-encoded sRGB caveat to color-math.md: @bjornlu/colorblind skips
  the sRGB->linear decode, so OKLab ΔE is directional, not exact
- Move renderSvg tests to generate-colorblind-svg.test.js and add a full
  buildSimulatedByVision -> renderSvg byte-stability case for the freshness gate
